+++ src/EventStore/EventStore.ts
@@ -5,12 +5,19 @@ import { web3 } from '../env'
 const contract  = require('truffle-contract') 
 const eventStoreArtifacts = require('../../build/contracts/EventStore')
 
-var ES = contract(eventStoreArtifacts)
-ES.setProvider(web3.currentProvider)
 
 import {
   Middleware
 } from './Middleware/Middleware'
+
+
+import {ReadModel} from '../ReadModel/ReadModel'
+
+export module EventStore{
+
+export const ReadModelGenerator = ReadModel
+export const ES = contract(eventStoreArtifacts)
+ES.setProvider(web3.currentProvider)
 
 /**
 * @param {TruffleContract} eventStore - a contract instance which is an Event Store
@@ -18,7 +25,7 @@ import {
 * @return {Promise<SOLIDITY_EVENT, Error>} json object representing a Solidity SOLIDITY_EVENT
 */
 export const readEvent = async (es, eventId) => {
-  return Middleware.readSolidityEventAsync(es, eventId)
+return Middleware.readSolidityEventAsync(es, eventId)
 }
 
 /**
@@ -27,13 +34,13 @@ export const readEvent = async (es, eventId) => {
 * @return {Promise<SOLIDITY_EVENT[], Error>} json objects representing SOLIDITY_EVENTs
 */
 export const readEvents = async (es, eventId = 0) => {
-  let currentEvent = await es.solidityEventCount()
-  let eventPromises = []
-  while (eventId < currentEvent) {
+let currentEvent = await es.solidityEventCount()
+let eventPromises = []
+while (eventId < currentEvent) {
     eventPromises.push(await readEvent(es, eventId))
     eventId++
-  }
-  return await Promise.all(eventPromises)
+}
+return await Promise.all(eventPromises)
 }
 
 /**
@@ -43,12 +50,12 @@ export const readEvents = async (es, eventId = 0) => {
 * @return {Promise<SOLIDITY_EVENT, Error>} json object representing the Solidity SOLIDITY_EVENT
 */
 export const writeEvent = async (es, transmuteEvent, fromAddress) => {
-  let meta = {
+let meta = {
     from: fromAddress,
     gas: 2000000
-  }
-  // console.log('writeSolidityEventAsync: ', transmuteEvent)
-  return await Middleware.writeSolidityEventAsync(es, meta, transmuteEvent)
+}
+// console.log('writeSolidityEventAsync: ', transmuteEvent)
+return await Middleware.writeSolidityEventAsync(es, meta, transmuteEvent)
 }
 
 /**
@@ -59,29 +66,14 @@ export const writeEvent = async (es, transmuteEvent, fromAddress) => {
 */
 export const writeEvents = async (es, eventArray, fromAddress) => {
 
-  let eventPromises = eventArray
+let eventPromises = eventArray
     .map((transmuteEvent) => {
-      return writeEvent(es, transmuteEvent, fromAddress)
+    return writeEvent(es, transmuteEvent, fromAddress)
     })
-  return await Promise.all(eventPromises)
+return await Promise.all(eventPromises)
     .then((newEvents) => {
-      return newEvents
+    return newEvents
     })
 }
 
-
-
-/**
-* @property {TruffleContract} ES - a truffle contract which is an EventStore
-* @property {readEvent} readEvent - read a single event by ID
-* @property {readEvents} readEvents - read a single event by ID, and all events after it
-* @property {writeEvent} writeEvent - write a single event
-* @property {writeEvents} writeEvents - write an array of events
-*/
-export const EventStore = {
-  ES,
-  readEvent,
-  readEvents,
-  writeEvent,
-  writeEvents
 }
