+++ src/TransmuteFramework.ts
@@ -1,0 +1,19 @@
+
+
+import { Persistence } from './Persistence/Persistence'
+import { EventTypes } from './EventTypes/EventTypes'
+import { Transactions } from './Transactions/Transactions'
+import { EventStore } from './EventStore/EventStore'
+
+import { ReadModel } from './ReadModel/ReadModel'
+
+export const TransmuteFramework = {
+  Persistence,
+  EventTypes,
+  Transactions,
+  EventStore,
+  ReadModel
+}
+
+export default TransmuteFramework
+
